refactor: data management to pure functions
